requirements.txt
discord.py[voice]
pynacl
discord.py
yt-dlp